+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## Unreleased
+### Added
+- Added support for verilator on windows
+
 ## [1.0.4] - 2019-11-09
 ### Added
 - Added logging to Output pane